Move the rest of the wrapper code from AppPickerWrapper to AppPicker component
